requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.25.0

# Authentication and Database
supabase==1.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi-jwt-auth==0.5.0

# Video processing
moviepy==1.0.3
opencv-python==4.8.1.78
whisper==1.1.10
pytesseract==0.3.10
pillow==10.0.1
scikit-learn==1.3.2

# Video downloading
pytube==15.0.0
yt-dlp==2023.11.16

# AI integration
openai==1.2.3
tiktoken==0.5.1

# Task management
networkx==3.1

# Web interface
jinja2==3.1.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-multipart==0.0.6
tqdm==4.66.1

# Export functionality
markdown==3.5.1
weasyprint==60.1
pyyaml==6.0.1

# Integrations
httpx==0.25.0
